requirements.txt
# Telegram Bot
aiogram==3.17.0
aiohttp==3.11.12

# Mini App
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database & config
python-dotenv==1.0.1

# Payments
# yookassa — используем прямые HTTP запросы через aiohttp (уже включён)

# Documents
docxtpl==0.20.1

# Charts (расписание)
matplotlib==3.10.0
pillow==11.1.0
numpy==2.2.2

# Validation
pydantic[email]==2.10.6
email-validator==2.2.0

# Utils
python-dateutil==2.9.0.post0
